// devtools-ui/src/api/model/moduleGraphNode.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Awilixify Devtools API
 * API for inspecting module graphs, traces, and provider impact
 * OpenAPI spec version: 1.0.0
 */
import type { DynamicModuleInfo } from "./dynamicModuleInfo";
import type { ModuleGraphNodeBase } from "./moduleGraphNodeBase";
import type { ModuleGraphNodeKind } from "./moduleGraphNodeKind";
import type { ModuleGraphNodeLifetimeTypes } from "./moduleGraphNodeLifetimeTypes";
import type { ModuleGraphNodeProviderAllowCircular } from "./moduleGraphNodeProviderAllowCircular";
import type { ModuleGraphNodeProviderDependencies } from "./moduleGraphNodeProviderDependencies";
import type { ModuleGraphNodeProviderEager } from "./moduleGraphNodeProviderEager";
import type { ModuleGraphNodeProviderInitAfter } from "./moduleGraphNodeProviderInitAfter";
import type { ModuleGraphRoute } from "./moduleGraphRoute";
import type { ModuleProviderImpact } from "./moduleProviderImpact";

export interface ModuleGraphNode {
	id: string;
	name: string;
	baseName?: string;
	dynamic?: DynamicModuleInfo;
	grouped: boolean;
	familyInstanceCount: number;
	instanceCount: number;
	kind: ModuleGraphNodeKind;
	dependencyCount: number;
	dependentCount: number;
	providers: string[];
	providerAllowCircular: ModuleGraphNodeProviderAllowCircular;
	providerDependencies: ModuleGraphNodeProviderDependencies;
	providerEager: ModuleGraphNodeProviderEager;
	providerInitAfter: ModuleGraphNodeProviderInitAfter;
	lifetimeTypes: ModuleGraphNodeLifetimeTypes;
	exports: string[];
	controllers: string[];
	queryHandlers: string[];
	commandHandlers: string[];
	queryPreHandlers: string[];
	queryPreHandlerExports: string[];
	commandPreHandlers: string[];
	commandPreHandlerExports: string[];
	interceptors: string[];
	interceptorExports: string[];
	initializers: string[];
	initializerExports: string[];
	routes: ModuleGraphRoute[];
	instances: ModuleGraphNodeBase[];
	impact: ModuleProviderImpact;
}